Cleaning up annotation scanning.
