+++ graylog2-server/src/main/java/org/graylog2/database/dbcatalog/DbEntitiesScanner.java
@@ -17,23 +17,27 @@
 package org.graylog2.database.dbcatalog;
 
 import com.google.common.base.Stopwatch;
-import org.graylog2.database.dbcatalog.impl.ClassGraphDbEntitiesScanningMethod;
-import org.graylog2.database.dbcatalog.impl.DbEntitiesScanningMethod;
+import io.github.classgraph.AnnotationParameterValueList;
+import io.github.classgraph.ClassGraph;
+import io.github.classgraph.ClassInfo;
+import io.github.classgraph.ClassInfoList;
+import io.github.classgraph.ScanResult;
+import jakarta.inject.Inject;
+import jakarta.inject.Provider;
+import org.graylog2.database.DbEntity;
 import org.graylog2.shared.plugins.ChainingClassLoader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import jakarta.inject.Inject;
-import jakarta.inject.Provider;
+import java.util.LinkedList;
+import java.util.List;
 
 public class DbEntitiesScanner implements Provider<DbEntitiesCatalog> {
-
     private static final Logger LOG = LoggerFactory.getLogger(DbEntitiesScanner.class);
 
     private final String[] packagesToScan;
     private final String[] packagesToExclude;
     private final ChainingClassLoader chainingClassLoader;
-    private final DbEntitiesScanningMethod dbEntitiesScanningMethod = new ClassGraphDbEntitiesScanningMethod();
 
     @SuppressWarnings("unused")
     @Inject
@@ -52,9 +56,48 @@ public class DbEntitiesScanner implements Provider<DbEntitiesCatalog> {
     @Override
     public DbEntitiesCatalog get() {
         final Stopwatch stopwatch = Stopwatch.createStarted();
-        final DbEntitiesCatalog catalog = dbEntitiesScanningMethod.scan(packagesToScan, packagesToExclude, chainingClassLoader);
+        final DbEntitiesCatalog catalog = scan(packagesToScan, packagesToExclude, chainingClassLoader);
         stopwatch.stop();
         LOG.info("{} entities have been scanned and added to DB Entity Catalog, it took {}", catalog.size(), stopwatch);
         return catalog;
+    }
+
+    public DbEntitiesCatalog scan(final String[] packagesToScan,
+                                  final String[] packagesToExclude,
+                                  final ChainingClassLoader chainingClassLoader) {
+        List<DbEntityCatalogEntry> dbEntities = new LinkedList<>();
+        ClassGraph classGraph = new ClassGraph()
+                .enableAnnotationInfo()
+                .acceptPackages(packagesToScan)
+                .rejectPackages(packagesToExclude)
+                .filterClasspathElements(classpathElementPathStr -> classpathElementPathStr.contains("graylog"))
+                .disableRuntimeInvisibleAnnotations();
+
+        if (chainingClassLoader != null) {
+            //Unfortunately, ClassGraph does not work correctly if provided with ChainingClassLoader as a whole
+            //You have to manually add all class loaders from ChainingClassLoader
+            for (ClassLoader cl : chainingClassLoader.getClassLoaders()) {
+                classGraph = classGraph.addClassLoader(cl);
+            }
+        }
+
+        try (ScanResult scanResult = classGraph.scan()) {
+            final String annotationName = DbEntity.class.getCanonicalName();
+            ClassInfoList classInfoList = scanResult.getClassesWithAnnotation(annotationName);
+            for (ClassInfo classInfo : classInfoList) {
+                final var annotations = classInfo.getAnnotationInfoRepeatable(annotationName);
+                for (final var annotationInfo : annotations) {
+                    AnnotationParameterValueList paramVals = annotationInfo.getParameterValues();
+                    dbEntities.add(new DbEntityCatalogEntry(
+                            paramVals.get("collection").getValue().toString(),
+                            paramVals.get("titleField").getValue().toString(),
+                            classInfo.loadClass(),
+                            paramVals.get("readPermission").getValue().toString()
+                    ));
+                }
+            }
+
+            return new DbEntitiesCatalog(dbEntities);
+        }
     }
 }
