--- graylog2-server/src/main/java/org/graylog2/database/dbcatalog/impl/DbEntitiesScanningMethod.java
@@ -1,27 +1,0 @@
-/*
- * Copyright (C) 2020 Graylog, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the Server Side Public License, version 1,
- * as published by MongoDB, Inc.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * Server Side Public License for more details.
- *
- * You should have received a copy of the Server Side Public License
- * along with this program. If not, see
- * <http://www.mongodb.com/licensing/server-side-public-license>.
- */
-package org.graylog2.database.dbcatalog.impl;
-
-import org.graylog2.database.dbcatalog.DbEntitiesCatalog;
-import org.graylog2.shared.plugins.ChainingClassLoader;
-
-public interface DbEntitiesScanningMethod {
-
-    DbEntitiesCatalog scan(final String[] packagesToScan,
-                           final String[] packagesToExclude,
-                           final ChainingClassLoader chainingClassLoader);
-}
